requirements.txt
Django==3.1.5
pillow==8.0.0
django-admin-interface==0.14.0
django-tinymce4-lite==1.8.0
django-cors-headers==3.4.0
django-admin-generator==2.2.0

# API
djangorestframework==3.12.2
django-filter==2.3.0